Implemented tests and MPM data subclass.
Created the MFEMMPMData subclass of MFEMData to handle particle
information. Implemented the MFEMDataTest and MFEMMPMTest classes.
Uncovered a bug in MFEMData where references could be returned to
uninitialized pointers in the unique_ptrs, so I added a check to make
sure that data was loaded and an exception thrown if not.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>